train/requirements-train.txt
-r ../requirements.txt
pandas==2.2.3
tensorflow-cpu==2.18.0
xgboost==2.1.1
